+++ plugin/buildSrc/src/main/kotlin/PluginDependencies.kt
@@ -1,6 +1,6 @@
 object PluginVersions {
     val kotlin = "1.3.0"
-    val ktlintPlugin = "6.2.1"
+    val ktlintPlugin = "6.3.0"
     val gradlePublishPlugin = "0.10.0"
     val androidPlugin = "3.2.0"
     val semver = "1.1.0"